Redesign serving/quantity UX
The new version makes it easier to add multiple items of
foods which come in a fixed size, like "3 yoghurts".

However, it makes it impossible to add a specific amount
(like 36 grams) of a food which has a serving size of
for example 120 grams. We will see if this is ever required.

Closes #17
Closes #14